Minor correction for TRX Report output
Modified to address an exception case which occurred when a user
'checked' a test entry in the Visual Studio Test Run viewer after the
results had been published to TFS. The cause of this was that the
correct id value for the main test list was not being assigned to each
test entry.

Altered the 'runUser' attribute in the 'TestRun' element to set the
correct domain and user in preference to the hard coded
"DOMAIN\UserName'' value.

The name of the TestList has been changed to 'StatLight Test List' so
that it is more easily identified when viewing the test results in the
Visual Studio Test Run viewer.

Updated the unit test to reflect changes to the TRX Report output and
altered the regex expression to support other time zones.